+++ CHANGES.md
@@ -12,6 +12,8 @@
 
 - Command `tag`: improve error and log messages by comparing the provided 
   commit with the commit correspondent to the provided tag (#226, @pitag-ha)
+- Error logs: when an external command fails, include its error message in
+  the error message posted by `dune-release` (#231, @pitag-ha)
 
 ### Deprecated